+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- feat(imports): add file import handler registration functionality. (`8d85a8f`)
+
 ## 1.0.63 - 2026-07-10
 
 - feat(footer-icon): add `activeStyle` prop with toolbar-matching `selection` variant.